requirements.txt
streamlit
pandas
psycopg2-binary
bcrypt
sqlalchemy
streamlit-calendar